refactor: content types to models
