+++ packages/prismic-dal/src/index.ts
@@ -1,21 +1,18 @@
 export * from "./gql";
 export type * from "./gql/graphql";
 
-/* content types */
-
-export * from "./content-types/peer";
-export * from "./content-types/technology";
-
 /* models */
 
 export * from "./models/case-study-body";
 export * from "./models/case-study-page";
 export * from "./models/home-page";
+export * from "./models/peer";
 export * from "./models/portfolio-page";
 export * from "./models/portfolio-page-featured";
 export * from "./models/sitemap-case-study";
 export * from "./models/sitemap-home-page";
 export * from "./models/sitemap-portfolio-page";
+export * from "./models/technology";
 
 /* queries */
