+++ packages/prismic-dal/src/models/technology/index.ts
